requirements.txt
Django==1.8.4
django-allauth==0.24.1
oauthlib==1.0.3
python-openid==2.2.5
requests==2.8.1
requests-oauthlib==0.5.0
wheel==0.24.0
